// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  TRAINER: 'TRAINER',
  SUPERVISOR: 'SUPERVISOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const RequestStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED',
  REVOKED: 'REVOKED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]